+++ wemake_python_styleguide/violations/best_practices.py
@@ -37,6 +37,7 @@ Summary
    BadMagicMethodViolation
    NestedImportViolation
    RedundantForElseViolation
+   RedundantFinallyViolation
 
 Comments
 --------
@@ -68,6 +69,7 @@ Design
 .. autoclass:: BadMagicMethodViolation
 .. autoclass:: NestedImportViolation
 .. autoclass:: RedundantForElseViolation
+.. autoclass:: RedundantFinallyViolation
 
 """
 
@@ -615,3 +617,37 @@ class RedundantForElseViolation(ASTViolation):
     #: Error message shown to the user.
     error_template = 'Found `else` in `for` loop with `break`'
     code = 436
+
+
+class RedundantFinallyViolation(ASTViolation):
+    """
+    Forbids to use ``finally`` in ``try`` block without ``except`` block.
+
+    Reasoning:
+        This rule will reduce complexity and improve readability.
+
+    Solution:
+        Refactor your ``try`` logic.
+        Replace the ``try-finally`` statement with a ``with`` statement.
+
+    Example::
+
+        # Correct:
+        with open("filename") as f:
+            f.write(...)
+
+        # Wrong:
+        try:
+            f = open("filename")
+            f.write(...)
+        finally:
+            f.close()
+
+    Note:
+        Returns Z437 as error code
+
+    """
+
+    #: Error message shown to the user.
+    error_template = 'Found `finally` in `try` block without `except`'
+    code = 437